LZMR-47 moved setCharactersOrgList to organizaion dao interface. implemented abstract method
