+++ Netbeans/SuperheroSightings/src/main/java/com/sg/superherosightings/controller/CharactersController.java
@@ -7,6 +7,8 @@ package com.sg.superherosightings.controller;
 
 import com.sg.superherosightings.dto.Characters;
 import com.sg.superherosightings.service.CharactersService;
+import com.sg.superherosightings.service.OrganizationsService;
+import java.util.List;
 import javax.inject.Inject;
 import javax.servlet.http.HttpServletRequest;
 import org.springframework.stereotype.Controller;
@@ -23,6 +25,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 public class CharactersController {
     
     CharactersService charactersService;
+    OrganizationsService organizationsService;
 
 @Inject
 public CharactersController(CharactersService charactersService) {
@@ -43,5 +46,43 @@ public CharactersController(CharactersService charactersService) {
         return "redirect:viewHeroes";
     }
 
+@RequestMapping(value = "/deleteHero", method = RequestMethod.GET)
+    public String deletHero(HttpServletRequest request, Model model) {
+        int heroID = Integer.parseInt(request.getParameter("characterId"));
+        charactersService.deleteCharacter(heroID);
+        return "redirect:viewHeroes";
+    }
     
+    
+@RequestMapping(value = {"/viewHeroes"}, method = RequestMethod.GET)
+    public String loadHeroes(HttpServletRequest request, Model model) {
+        List<Characters> allHeroes = charactersService.getAllHeroes();
+        organizationsService.setCharactersOrgList(allHeroes);
+        model.addAttribute("heroes", allHeroes);
+
+        String display = (request.getParameter("viewType"));
+        model.addAttribute("display", display);
+
+        try {
+            int id = Integer.parseInt(request.getParameter("characterId"));
+            Characters hero = charactersService.getCharacterById(id);
+            model.addAttribute("hero", hero);
+
+        } catch (Exception e) {
+
+        }
+
+        return "heroes";
+    }
+
+
+
+
+
+
+
+
+
+
+
 }
