+++ Netbeans/SuperheroSightings/src/main/java/com/sg/superherosightings/service/OrganizationsService.java
@@ -29,5 +29,7 @@ public interface OrganizationsService {
     
     public List<Organization> getOrganizationByLocationId(int locationId);
     
+    public void setCharactersOrgList(List<Characters> temp);
+
     
 }
